# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /restaurants-mean:
    x-swagger-router-controller: restaurants.mean
    get:
      summary: Displays all of the restaurants available from Mongo
      description: Displays all of the restaurants available.  These restaurants are fantastic.
      operationId: indexMean
      responses:
        200:
          description: An array of restaurants
          schema:
            $ref: "#/definitions/Restaurants"
    post:
      summary: Create a new restaurant entry
      description: Create a new restaurant entry
      operationId: createMean
      parameters:
        - name: restaurant
          in: body
          required: true
          description: Restaurant Object
          schema:
            $ref: "#/definitions/Restaurant"         
      responses:
        201:
          description: OK
          schema:
            $ref: "#/definitions/Restaurant"
...
definitions:
  Restaurants:
    items:
      $ref: "#/definitions/Restaurant"
  Restaurant:
    properties:
      name:
        type: string
      address:
        type: string
...